fix(mobile): validate message permalink chips strictly
